chore: rename ms/metrics to ms/mono
