+++ env.sh.dist
@@ -22,8 +22,8 @@ export mono_nats_addr_port="4222"
 #   <OTHER>                 - used by 3rd-party libraries/frameworks/tools
 export GO_TEST_TIME_FACTOR="1.0" # Increase if tests fail because of slow CPU.
 export MONO_EXAMPLE_ADDR_PORT="17001"
-export MONO_EXAMPLE_METRICS_ADDR_PORT="16001"
-export MONO_METRICS_ADDR_PORT="16000"
+export MONO_EXAMPLE_METRICS_ADDR_PORT="17002"
+export MONO_MONO_ADDR_PORT="17000"
 export MONO_X_MYSQL_ADDR_HOST="localhost"
 export MONO_X_MYSQL_ADDR_PORT="${mono_mysql_addr_port}"
 export MONO_X_NATS_ADDR_URLS="nats://localhost:${mono_nats_addr_port}"
