+++ internal/config/config.go
@@ -20,33 +20,28 @@ const EnvPrefix = "MONO_"
 
 // Shared contains configurable values shared by microservices.
 type Shared struct {
+	AddrHost               appcfg.NotEmptyString `env:"ADDR_HOST"`
+	AddrHostInt            appcfg.NotEmptyString `env:"ADDR_HOST_INT"`
 	ExampleAddrPort        appcfg.Port           `env:"EXAMPLE_ADDR_PORT"`
 	ExampleMetricsAddrPort appcfg.Port           `env:"EXAMPLE_METRICS_ADDR_PORT"`
-	MonoAddrHost           appcfg.NotEmptyString `env:"MONO_ADDR_HOST"`
-	MonoMetricsAddrHost    appcfg.NotEmptyString `env:"MONO_METRICS_ADDR_HOST"`
 	XMySQLAddrHost         appcfg.NotEmptyString `env:"X_MYSQL_ADDR_HOST"`
 	XMySQLAddrPort         appcfg.Port           `env:"X_MYSQL_ADDR_PORT"`
 	XNATSAddrUrls          appcfg.NotEmptyString `env:"X_NATS_ADDR_URLS"`
 	XSTANClusterID         appcfg.NotEmptyString `env:"X_STAN_CLUSTER_ID"`
 }
 
-// Default ports for metrics.
+// Default ports.
 const (
-	MetricsPort = 16000 + iota // XXX Used in ms/metrics/service.go.
+	MonoPort = 17000 + iota
+	ExamplePort
 	ExampleMetricsPort
 )
 
-// Default ports.
-const (
-	_ = 17000 + iota
-	ExamplePort
-)
-
 var shared = &Shared{ //nolint:gochecknoglobals // Config is global anyway.
+	AddrHost:               appcfg.MustNotEmptyString(def.Hostname),
+	AddrHostInt:            appcfg.MustNotEmptyString(def.Hostname),
 	ExampleAddrPort:        appcfg.MustPort(strconv.Itoa(ExamplePort)),
 	ExampleMetricsAddrPort: appcfg.MustPort(strconv.Itoa(ExampleMetricsPort)),
-	MonoAddrHost:           appcfg.MustNotEmptyString(def.Hostname),
-	MonoMetricsAddrHost:    appcfg.MustNotEmptyString(def.Hostname),
 	XMySQLAddrPort:         appcfg.MustPort("3306"),
 }
